Seal TupledFunction to only have instances in internal
